Add support for exporting and importing comment metadata
When pages in the CommentStreams namespace are imported/exported
(using Special:Import and Special:Export), comment metadata is
imported/exported as well. If a corresponding user to the commenter
exists on the import wiki, the comments/replies will be attributed
to that user; otherwise, the comments/replies will be anonymous.

This patch does *not* automatically export comments on a page when that
page is exported. That will likely need a core patch to allow extra
pages to be added to the export list by an extension in a hook.

Bug: T292377
Change-Id: Ic16cd254bc946fd93b03203898ac2422af92d003