Support for multiple comments blocks in a page
Added support for multiple comments block in one page. To create a
comment block add a id param with some unique string (page-level)
Eg: <comment-streams id="unique_id"/>

Normal functionality of having <comment-streams/> would work too.
Internally this is done by assigning a new field cst_id for the parent
comments. The value for this field is set as a md5 of the string
provided by the user and the default value for old comments is "cs-comments".

Bumped version to 5.0

Bug: T246487
Change-Id: I78d92a61560363b40d184cf41d28bdbcca51e245