Convert to OOUI
This patch also removes the <comment-streams location="footer"/>
feature. That feature had become inoperable at some point. Now, the
default comment block will always appear in the page footer. Named
comment blocks can be placed anywhere in the content area of the page.
This is irrespective of whether comments are enabled on a page with
$wgCommentStreamsAllowedNamespaces or by including <comment-streams />
(with or without a comment block id provided).

Change-Id: Ie23e4159a9f81ea0d8d7c1325c1f594ec98ee1ef