Stop using md5 for comment block ids.
The cst_id column of the cs_comment_data table currently holds the MD5
hash of the comment block name prepended by 'csc_' with the exception of
the default comment block, which is designated by 'cs-comments' in the
cst_id column. This makes it extremely difficult to maintain comment
blocks. If a comment block gets renamed in wikitext, there does not
currently exist a tool to identify which rows need to get adjusted.

This patch replaces the value in the cst_id column with null for the
default comment block and with the name of the comment block (unhashed)
for named comment blocks. It no longer uses the comment block name as an
id for the element in the DOM. It includes three maintenance scripts.

maintenance/NullDefaultCommentBlock is run on upgrade by
maintenance/update.php to null out the cst_id column for rows that
previously were set to 'cs-comments' (the default comment block).

maintenance/listCommentBlocks lists all of the comment blocks either
across the entire wiki or, with the 'pages' option, sorted by the page
they appear on.

maintenance/migrateCommentBlock takes the unhashed name of a comment
block and either replaces cst_id with the unhashed name for all rows
that currently hold 'csc_' followed by the hashed named - or -
takes both a from name and a to name to rename any comment block.

Change-Id: I058f342de202b16c1371157ea53e1d356c7b6ce8