Fix associatedId not always corresponding to the page comments are shown for
Commenting on a talk page causes the comment to get lost as it will
be associated with the talk page, but the comments displayed will
only be the ones associated to the subject page.
This is fixed by passing the ID of the associated page to the module
and using it instead of wgArticleId (which would be the ID of the
talk of and not the one of the associated subject page).

Bug: T397582
Change-Id: I55ac64b0309df40495b019fb652d41996e327bc9