Add a parent id parameter to EchoEvent extra for reply notifications
In locateUsersWatchingComment(), comment_id is used to retrieve
users watching a comment. However, when a reply is made, the reply
id is passed to this parameter, causing the method to try to find users
that are watching the reply, which returns no one as replies cannot be
watched.

comment_id is used to get the url of any comment including replies,
so for reply notifications, it is suitable to pass the parent comment id
to a new parameter instead of changing comment_id.

Bug: T411809
Change-Id: Ie550c35c8224a7a926775f563fd9963f2a902211